feat: split FAQ + add information | feat 2 : incremental name
