+++ client/src/app/(metadraft)/(validator)/metadata-structure/how-to/index.tsx
@@ -1,0 +1,27 @@
+import { Typography } from "~/components/typography";
+import { Accordion } from "~/components/ui/accordion";
+
+import WhatAreType from "./what-are-type";
+import WhatAreString from "./what-are-string";
+import WhatAreNumber from "./what-are-number";
+import WhatAreObject from "./what-are-object";
+import WhatAreArray from "./what-are-array";
+import DefaultIsValue from "./default-is-value";
+import IncrementalName from "./incremental-name";
+
+export default function HowToCreateMetadataSchema() {
+  return (
+    <div className="flex flex-col gap-2 rounded-xl bg-secondary p-4">
+      <Typography as="h3">FAQ</Typography>    
+      <Accordion type="single" collapsible>
+        <WhatAreType />
+        <DefaultIsValue />
+        <IncrementalName />
+        <WhatAreString />
+        <WhatAreNumber />
+        <WhatAreObject />
+        <WhatAreArray />
+      </Accordion>
+    </div>
+  );
+}
