+++ client/src/app/(metadraft)/(validator)/metadata-structure/how-to/what-are-number.tsx
@@ -1,0 +1,28 @@
+import { Typography } from "~/components/typography";
+import {
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "~/components/ui/accordion";
+import ColumnInnerContent from "./column-inner-content";
+
+export default function WhatAreNumber() {
+  return (
+    <AccordionItem value="value-number" className="border-b border-border/10">
+      <AccordionTrigger>
+        <Typography as="code" className="text-lg">
+          How to use number value ?
+        </Typography>
+      </AccordionTrigger>
+      <AccordionContent>
+        <ColumnInnerContent>
+          <Typography as="code">{`Simply enter 0 as the value for any key that you want to be treated as a number.`}</Typography>
+          <Typography
+            as="code"
+            className="ml-4 text-white/50"
+          >{`Example: { age: 0 }`}</Typography>
+        </ColumnInnerContent>
+      </AccordionContent>
+    </AccordionItem>
+  );
+}
